refactor: Enhance channel structure and factor out common abstractions
